refactor: extract table layout, add line/state mapping and scroll indicators
